added action button to bottom screen
